Add lexer implementation
The implementation relies on the function open_memstream,
which is a part of the POSIX C library spec. It basically
associates a file stream to a dynamic memory buffer where
allocation is managed by the C library implementation.

The token is recognized and is written to this file
stream using fwrite, kindof simulating the push_back method
on std::vector in C++.